2023 Day 21: Step Counter - part 2
Got some help from reddit to understand what to do with the detected 131 cycle pattern.

Seeing that the desired 26,501,365 steps is (202,300 * 131) + 65, assuming we want a formula where 202,300 is x, then we take the first 3 points of x = 0, x = 1, and x = 2:

- (0 * 131) + 65 = 65
- (1 * 131) + 65 = 196
- (2 * 131) + 65 = 327

Get the total number of locations after each of those iterations, and back into creating a quadratic formula based on those 3 points.

With that quadratic formula of `15397x^2 + 15495x + 3893` , plug in x of 202,300 to get the answer: `630129824772393`

https://www.reddit.com/r/adventofcode/comments/18nevo3/comment/keao6r4/
https://socratic.org/questions/given-three-points-0-3-1-4-2-9-how-do-you-write-a-quadratic-function-in-standard